@ #11 | fix bug failed to sync created files from Linux host
