+++ lib/teracy-dev-v05-compat.rb
@@ -1,6 +1,6 @@
 require 'teracy-dev'
 
-require_relative 'teracy-dev-v05-compat/config/gatling_rsync_recovery'
+require_relative 'teracy-dev-v05-compat/config/rsync_recovery'
 require_relative 'teracy-dev-v05-compat/processors/multiple_essential_support'
 require_relative 'teracy-dev-v05-compat/processors/provisioner_id_deprecated_removal'
 
@@ -10,7 +10,7 @@ module TeracyDevV05Compat
     TeracyDev.register_processor(Processors::MultipleEssentialSupport.new)
     TeracyDev.register_processor(Processors::ProvisionerIdDeprecatedRemoval.new)
 
-    TeracyDev.register_configurator(TeracyDevV05Compat::Config::GatlingRsyncRecovery.new)
+    TeracyDev.register_configurator(TeracyDevV05Compat::Config::RsyncRecovery.new)
   end
 
 end
